requirements.txt
psutil
Pillow
opencv-python
numpy
pandas
requests
uiautomator2